Moved the organization pipelines to it's own component
